Atualização das funções
Foi realizada uma correção no caminho do diretório das funções e adaptações para a conexão com o servidor MySQL.
