+++ Functions/ConexaoMySQL.py
@@ -1,23 +1,20 @@
 import mysql
-from mysql import errorcode
+#from mysql.connector import errorcode
 
 def Conecta_MySQL():
 	try:
-		db_connection = mysql.connector.connect(host = '127.0.0.1', user = 'root', password = '', database = 'Resultado_testes'); 
+		db_connection = mysql.connector.connect(host = '127.0.0.1', user = 'root', password = '1234', database = 'resultado_testes'); 
 		#aqui é realizado a tentativa de conectar ao host local do computador
 		print("Conexão com o Banco de Dados feita com sucesso!");
 		return 1;
-	except mysql.connector.Error as error:
-		if error.errno == errorcode.ER_BAD_DB_ERROR:
-			print("O banco de dados não existe!"); #caso isso aconteça, é porque o banco de dados não existe
-		else:
-			print("O servidor do banco de dados está indisponível no momento!"); #o servidor está offline;
-			return 0;
+	except Exception as error:
+		print("O servidor do banco de dados está indisponível no momento!"); #o servidor está offline;
+		return 0;
 	else:
 	    db_connection.close();
 
 def Get_Dados_MySQL():
-	db_connection = mysql.connector.connect(host = '127.0.0.1', user = 'root', password = '', database = 'Resultado_testes');
+	db_connection = mysql.connector.connect(host = '127.0.0.1', user = 'root', password = '1234', database = 'Resultado_testes');
 	cursor = db_connection.cursor();
 	sql = ("SELECT AVG(Download), AVG(Upload), AVG(Ping), AVG(Jitter) FROM Resultados");
 	cursor.execute(sql);
